database/pyproject.toml
[tool.poetry]
name = "my_project"
version = "0.1.0"
description = ""
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
pydantic = "^2.8.2"
python-docx = "^0.8.11"
requests = "^2.26.0"
chromadb = "0.5.3"
pdfplumber = "^0.5.28"
sentence-transformers = "^3.0.1"
python-dotenv = "^1.0.1"
aiofiles = "^24.1.0"
langchain = "^0.2.16"
langchain-experimental = "^0.0.65"
langchain-huggingface = "^0.0.3"
marker-pdf = "^0.2.17"
pathlib = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
